app configuration logic moved to a new file
